pyproject.toml
[tool.poetry]
name = "strawberry-migration-test"
version = "0.1.0"
description = ""
authors = ["Alteian <Alteian@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.2"
strawberry-graphql = "^0.189.0"
strawberry-graphql-django = "^0.9.5"
strawberry-django-plus = "^3.0.2"
pyjwt = "^2.7.0"
cryptography = "^41.0.1"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.6"
gunicorn = "^20.1.0"
uvicorn = "^0.22.0"

[tool.poetry.group.dev.dependencies]
django-pytest = "^0.2.0"
ruff = "^0.0.275"
black = "^23.3.0"
django-extensions = "^3.2.3"
django-debug-toolbar = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]  # noqa: B018, W291, F821

select = ["ALL"]
ignore = [
    "EXE002",
    # https://github.com/astral-sh/ruff/pull/4427
    # equivalent to keep-runtime-typing
    "UP006",
    "UP007",

    "TID252",
    # we use asserts in tests and to hint mypy
    "S101",
    "S102",
    "S104",
    "S324",
    # maybe we can enable this in future
    # we'd want to have consistent docstrings in future
    "D",
    "ANN101", # missing annotation for self?
    # definitely enable these, maybe not in tests
    "ANN102",
    "ANN202",
    "ANN204",
    "ANN205",
    "ANN401",
    "PGH003",
    "PGH004",
    "RET504",
    "RET505",
    "RET506",
    "RET507",
    "RET503",
    "BLE001",
    "B008",
    "N811",
    "N804",
    "N818",
    # Variable `T` in function should be lowercase
    # this seems a potential bug or opportunity for improvement in ruff
    "N806",

    # first argument should named self (found in tests)
    "N805",

    "N815",

    # shadowing builtins
    "A001",
    "A002",
    "A003",

    "ARG001",
    "ARG002",
    "ARG003",
    "ARG004",
    "ARG005",
    "FBT001",
    "FBT002",
    "FBT003",

    "PT001",
    "PT023",

    # enable these, we have some in tests
    "B006",
    "PT004",
    "PT007",
    "PT011",
    "PT012",
    "PT015",
    "PT017",
    "C414",
    "N802",

    "SIM117",
    "SIM102",

    "F841",
    "B027",
    "B905",
    "ISC001",

    # same?
    "S105",
    "S106",

    "DTZ003",
    "DTZ005",

    "RSE102",
    "SLF001",

    # in tests
    "DTZ001",

    "EM101",
    "EM102",
    "EM103",

    "B904",
    "B019",

    "N801",
    "N807",

    # pandas
    "PD",

    # code complexity
    "C",
    "C901",

    # trailing commas
    "COM812",

    "PLR",
    "INP",
    "TRY",
    "SIM300",
    "SIM114",

    "DJ008",
    "TD002",
    "TD003",
    "FIX001",
    "FIX002",
    "FA100",
]

fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]  # noqa: E501, W291
unfixable = []

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "venv",
    "*.txt",
    "*.md"
]

line-length = 79  # noqa: W291, F821

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"  # noqa: W291, F821

[tool.ruff.isort]  # noqa: B018, W291, F821
known-third-party = ["django", "rest_framework"]  # noqa: W291, F821
extra-standard-library = ["typing_extensions"]  # noqa: W291, F821

[tool.ruff.mccabe]  # noqa: B018, W291, F821
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10  # noqa: W291, F821

[tool.ruff.flake8-quotes]  # noqa: W291, F821
docstring-quotes = "double"  # noqa: W291, F821

[tool.ruff.pydocstyle]  # noqa: B018, W291, F821
convention = "google"

[tool.black]  # noqa: B018, W291, F821
line-length = 79  # noqa: W291, F821